Fix issue #72

+++ flexviews/procs/disable.sql
@@ -51,6 +51,7 @@ BEGIN
   -- DECLARE v_mview_enabled tinyint(1);
   DECLARE v_mview_name TEXT character set utf8;
   DECLARE v_mview_schema TEXT character set utf8;
+  DECLARE v_mview_refresh_type TEXT character set utf8;
   DECLARE v_mview_enabled INT;
 
   DECLARE v_child_mview_id INT;
@@ -68,54 +69,57 @@ BEGIN
 
   SELECT mview_name, 
          mview_schema,
-	 mview_enabled
+	 mview_enabled, 
+         mview_refresh_type
     INTO v_mview_name, 
          v_mview_schema,
-         v_mview_enabled
+         v_mview_enabled,
+         v_mview_refresh_type
     FROM flexviews.mview
    WHERE flexviews.mview.mview_id = v_mview_id;
 
-   SELECT 'This procedure is deprecated.  Please use flexviews.DROP() to remove a view or flexviews.INVALIDATE() to mark it as invalid' as `WARNING` from dual;
+   IF v_mview_refresh_type != 'COMPLETE' THEN 
+     SELECT 'This procedure is deprecated.  Please use flexviews.DROP() to remove a view or flexviews.INVALIDATE() to mark it as invalid' as `WARNING` from dual;
 
-   IF v_mview_id IS NULL THEN
-     CALL flexviews.signal('The specified materialized view does not exist (NOTHING WAS DROPPED)');
+     IF v_mview_id IS NULL THEN
+       CALL flexviews.signal('The specified materialized view does not exist (NOTHING WAS DROPPED)');
+     END IF;
+
+     IF v_mview_enabled = FALSE OR v_mview_enabled is null THEN
+       CALL flexviews.signal('This materialized view is already disabled (NOTHING WAS DROPPED)');
+     END IF;
+
+     SELECT mview_id
+       INTO v_child_mview_id
+       FROM flexviews.mview
+      WHERE flexviews.mview.parent_mview_id = v_mview_id;
+
+     IF v_child_mview_id IS NOT NULL THEN
+       CALL flexviews.disable(v_child_mview_id);
+     END IF;
+
+     -- This will be committed by the DROP
+     UPDATE flexviews.mview
+        SET mview_last_refresh = NULL,
+            mview_enabled = FALSE
+      WHERE flexviews.mview.mview_id = v_mview_id;
+
+     SET @v_sql = CONCAT('DROP TABLE IF EXISTS ', v_mview_schema, '.', v_mview_name);
+     PREPARE drop_stmt FROM @v_sql; 
+     EXECUTE drop_stmt;
+
+     SELECT 'The materialized view was dropped sucessfully' as `MESSAGE` from dual;
+
+     SET @v_sql = CONCAT('DROP TABLE IF EXISTS ', v_mview_schema, '.', v_mview_name, '_delta');
+     PREPARE drop_stmt FROM @v_sql; 
+     EXECUTE drop_stmt;
+
+     SET @v_sql := NULL;
+     DEALLOCATE PREPARE drop_stmt;
+     SELECT 'The materialized view delta table was dropped sucessfully' as `MESSAGE` from dual;
+
+     SELECT 'View is now disabled.' as `MESSAGE` from dual;
    END IF;
-
-   IF v_mview_enabled = FALSE OR v_mview_enabled is null THEN
-     CALL flexviews.signal('This materialized view is already disabled (NOTHING WAS DROPPED)');
-   END IF;
-
-   SELECT mview_id
-     INTO v_child_mview_id
-     FROM flexviews.mview
-    WHERE flexviews.mview.parent_mview_id = v_mview_id;
-
-   IF v_child_mview_id IS NOT NULL THEN
-     CALL flexviews.disable(v_child_mview_id);
-   END IF;
-
-   -- This will be committed by the DROP
-   UPDATE flexviews.mview
-      SET mview_last_refresh = NULL,
-          mview_enabled = FALSE
-    WHERE flexviews.mview.mview_id = v_mview_id;
-
-   SET @v_sql = CONCAT('DROP TABLE IF EXISTS ', v_mview_schema, '.', v_mview_name);
-   PREPARE drop_stmt FROM @v_sql; 
-   EXECUTE drop_stmt;
-
-   SELECT 'The materialized view was dropped sucessfully' as `MESSAGE` from dual;
-
-   SET @v_sql = CONCAT('DROP TABLE IF EXISTS ', v_mview_schema, '.', v_mview_name, '_delta');
-   PREPARE drop_stmt FROM @v_sql; 
-   EXECUTE drop_stmt;
-
-   SET @v_sql := NULL;
-   DEALLOCATE PREPARE drop_stmt;
-   SELECT 'The materialized view delta table was dropped sucessfully' as `MESSAGE` from dual;
-
-   SELECT 'View is now disabled.' as `MESSAGE` from dual;
- 
    -- restore SESSION max_sp_recursion_depth
    SET max_sp_recursion_depth := bkp_max_sp_recursion_depth;
 END ;;
